th-2: Add time norms

+++ TH.Dal/Entities/TimeNorm.cs
@@ -5,5 +5,5 @@ public class TimeNorm
 	public Guid Id { get; set; }
 	public string Code { get; set; } = String.Empty;
 	public string Name { get; set; } = String.Empty;
-	public int Norm { get; set; }
+	public double Norm { get; set; }
 }
